requirements.txt
python-socketio==4.3.1
aiohttp==3.6.1